refactor : 기존 API에 로그인 반영 (3/3)

+++ src/main/java/mashup/backend/tich/faq/controller/AdminFaqController.java
@@ -2,19 +2,17 @@ package mashup.backend.tich.faq.controller;
 
 import io.swagger.annotations.ApiOperation;
 import lombok.RequiredArgsConstructor;
+import mashup.backend.tich.exception.NoAccessException;
 import mashup.backend.tich.faq.dto.FaqResponseDto;
 import mashup.backend.tich.faq.dto.FaqSaveRequestDto;
 import mashup.backend.tich.faq.dto.FaqUpdateRequestDto;
 import mashup.backend.tich.faq.service.AdminFaqService;
 import mashup.backend.tich.faq.service.FaqService;
-import mashup.backend.tich.user.domain.Role;
-import mashup.backend.tich.user.domain.User;
-import mashup.backend.tich.user.domain.UserRepository;
+import mashup.backend.tich.user.service.AdminUserService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.persistence.NoResultException;
 import java.util.List;
 
 @RequiredArgsConstructor
@@ -23,12 +21,14 @@ import java.util.List;
 public class AdminFaqController {
 
     private final AdminFaqService adminFaqService;
+    private final AdminUserService adminUserService;
     private final FaqService faqService;
-    private final UserRepository userRepository; /* 삭제 예정(관리자 확인용) */
 
     @ApiOperation("FAQ 목록 조회")
     @GetMapping
-    public ResponseEntity<List<FaqResponseDto>> showFaqs() {
+    public ResponseEntity<List<FaqResponseDto>> showFaqs(@RequestHeader("TICH-TOKEN") String token) {
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
+
         List<FaqResponseDto> faqResponseDto = faqService.showFaqs();
 
         return ResponseEntity.status(HttpStatus.OK).body(faqResponseDto);
@@ -36,7 +36,10 @@ public class AdminFaqController {
 
     @ApiOperation("FAQ 상세 조회")
     @GetMapping("/{faqId}")
-    public ResponseEntity<FaqResponseDto> showFaq(@PathVariable Long faqId) {
+    public ResponseEntity<FaqResponseDto> showFaq(@RequestHeader("TICH-TOKEN") String token,
+                                                  @PathVariable Long faqId) {
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
+
         FaqResponseDto faqResponseDto = faqService.showFaq(faqId);
 
         return ResponseEntity.status(HttpStatus.OK).body(faqResponseDto);
@@ -44,11 +47,9 @@ public class AdminFaqController {
 
     @ApiOperation("FAQ 등록")
     @PostMapping
-    public ResponseEntity<FaqResponseDto> saveFaq(@RequestHeader String accessToken,
+    public ResponseEntity<FaqResponseDto> saveFaq(@RequestHeader("TICH-TOKEN") String token,
                                                   @RequestBody FaqSaveRequestDto requestDto) {
-        // 임시 코드 : 추후 수정
-        User user = makeTempUser();
-        // ToDo : user check (accessToken)
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         FaqResponseDto faqResponseDto = adminFaqService.saveFaq(requestDto);
 
@@ -57,11 +58,9 @@ public class AdminFaqController {
 
     @ApiOperation("FAQ 수정")
     @PutMapping
-    public ResponseEntity<FaqResponseDto> updateFaq(@RequestHeader String accessToken,
+    public ResponseEntity<FaqResponseDto> updateFaq(@RequestHeader("TICH-TOKEN") String token,
                                                     @RequestBody FaqUpdateRequestDto requestDto) {
-        // 임시 코드 : 추후 수정
-        User user = makeTempUser();
-        // ToDo : user check (accessToken)
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         FaqResponseDto faqResponseDto = adminFaqService.updateFaq(requestDto);
 
@@ -70,31 +69,12 @@ public class AdminFaqController {
 
     @ApiOperation("FAQ 삭제")
     @DeleteMapping("/{faqId}")
-    public ResponseEntity<Void> deleteFaq(@RequestHeader String accessToken,
+    public ResponseEntity<Void> deleteFaq(@RequestHeader("TICH-TOKEN") String token,
                                           @PathVariable Long faqId) {
-        // 임시 코드 : 추후 수정
-        User user = makeTempUser();
-        // ToDo : user check (accessToken)
+        if (!adminUserService.adminByToken(token)) throw new NoAccessException("Admin only");
 
         adminFaqService.deleteFaq(faqId);
 
         return ResponseEntity.status(HttpStatus.OK).build();
-    }
-
-    /* 임시 코드 : 삭제 예정 */
-    private User makeTempUser() {
-        User user;
-        try {
-            user = userRepository.findByEmail("admin")
-                    .orElseThrow(() -> new NoResultException());
-        } catch (NoResultException e) {
-            user = userRepository.save(User.builder()
-                    .name("관리자")
-                    .email("admin")
-                    .picture("temp")
-                    .role(Role.ADMIN)
-                    .build());
-        }
-        return user;
     }
 }
